+++ CHANGES.md
@@ -1,4 +1,7 @@
 # CHANGES
+## 2.4.0
+Added continuous differentiation of cells between compartments.
+Now export the population trajectory of individual compartments.
 ## 2.3.0
 Added death rate to trajectory.
 ## 2.2.7